move internal test to internal_test.go
